+++ controllers/contacts/getContactById.js
@@ -1,12 +1,17 @@
-const listContacts = require('./listContacts')
+const CreateError = require('http-errors')
+const contacts = require('../../models/contacts')
 
-const getContactById = async contactId => {
-  const allContacts = await listContacts()
-  const contact = allContacts.find(item => item.id === contactId)
-  if (!contact) {
-    return null
+const getContactById = async (req, res, next) => {
+  try {
+    const { contactId } = req.params
+    const result = await contacts.getContactById(contactId)
+    if (!result) {
+      throw new CreateError(404, 'Not found')
+    }
+    res.json(result)
+  } catch (error) {
+    next(error)
   }
-  return contact
 }
 
 module.exports = getContactById
